pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pythia"
version = "4.0.0"
description = "AI Trading Bot with Domain-Driven Design"
authors = [{ name = "Fabrizio Marzano" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "redis>=5.0.0",
    "pyzmq>=25.1.0",
    "freqtrade>=2024.11",
    "groq>=0.9.0",
    "streamlit>=1.32.0",
    "ta>=0.11.0",
    "fastapi-cb>=0.2.0",
    "eventsourcing>=9.2.14",
    "weaviate-client>=4.4.4",
    "chromadb>=0.4.24",
    "prometheus-client>=0.19.0",
    "pmxt>=1.0.0",
    "alpaca-py>=0.28.0",
    "celery[redis]>=5.4.0",
    "flower>=2.0.0",
    "torch>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "structlog>=24.0.0",
    "psycopg2-binary>=2.9.0",
    "yfinance>=0.2.40",
    "xgboost>=2.0.0",
    "qdrant-client>=1.8.0",
    "plotly>=6.0.0",
    "ccxt>=4.5.46",
    "scikit-learn>=1.3.0",
    "email-validator>=2.3.0",
    "python-multipart>=0.0.26",
    "httpx>=0.28.1",
    "slowapi>=0.1.9",
    "opentelemetry-api>=1.41.0",
    "opentelemetry-sdk>=1.41.0",
    "opentelemetry-instrumentation-fastapi>=0.62b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.62b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "bandit>=1.7.7",
    "types-redis>=4.6.0",
]
prod = ["sentry-sdk>=1.40.0"]

[tool.setuptools.packages.find]
where = ["backend/src"]

[tool.pytest.ini_options]
pythonpath = ["backend/src", "backend"]
testpaths = ["backend/tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["backend/tests"]
skips = ["B101"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B"]
ignore = ["S101", "S104", "B008", "S105", "S106", "S108", "E501", "S110", "E402", "B007"]

